Change Id from Integer -> Long Function Add and Edit row AbstractListView, AbstractDetailView is basically finished. => We should not add field id in AbstractDetailView

+++ ssm/ssm-core/src/main/java/com/hbsoft/ssm/entity/AbstractBaseIdObject.java
@@ -15,16 +15,16 @@ import javax.persistence.MappedSuperclass;
 
 @MappedSuperclass
 public abstract class AbstractBaseIdObject {
-    private Integer id;
+    private Long id;
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
